+++ Sample/VPLviaSVOGI/Source/Private/Renderers/Visualizer.cpp
@@ -454,9 +454,9 @@ void Visualizer::Initialize(GPUDevice* device, Voxelizer* voxelizer,
         svoCubeVertices.push_back(vec3(1.0f, -1.0f, 1.0f));
         svoCubeVertices.push_back(vec3(1.0f, -1.0f, -1.0f));
         svoCubeVertices.push_back(vec3(-1.0f, -1.0f, -1.0f));
-        std::vector<unsigned short> svoCubeIndices;
+        std::vector<unsigned int> svoCubeIndices;
         svoCubeIndices.reserve(24);
-        static unsigned short indices[] = { 0, 1, 2, 3,
+        static unsigned int indices[] = { 0, 1, 2, 3,
             0, 4, 5, 1,
             0, 3, 7, 4,
             2, 6, 7, 3,
